Error handling now works for the ping command.
Instead of pulling the error exception inside the function we can use the event handling that ext has to only be checked when an actual error is called. In the documentation at ext the order of error and ctx was switched so it took a long time to find this. I also changed the format of some of our longer strings so they fit on half a screen because it looks cleaner. Pretty sure it's part of pep8 or something as well but we can worry about that later.