Ping Error Handles User Not Found in Directory
Ping did not handle a user not being in the dictionary and put a cooldown on the user. We shouldn't punish the user from the command if it doesn't execute the command.

Found out how to use reset_cooldown. This can be abused to reset someone's cooldown if someone already has one on them though. We'll have to implement some kind of check for that. Maybe we can change this so an error is sent at command call instead.