Properly Handle No Arguments Provided.
I was wondering why it got past the discord.Member formatting and then realized that it looks like we set it to None after the formatting. That would explain why we get passed NoneType.

We'll avoid manually raising our own exception and then we can add it back into the error handling function. Added it to the BadArgument condition because that's really where it fits.