Fixed "reset_cooldown" error. Mentions are now Name. Strings are now F-Strings (#6)
* Initial

Implementing cogs so we are able to work in separate files and we can easily work on separate functions.

Doing it how we were currently doing it, SahibBot.py would have become a huge file and probably not very managable.

* Moved Error and Ping to their own cogs.

All of the ping and error handling are now in their own files and it works properly.

* Cleanup and made Twilio Live

We're able to comment out and unload extensions now so lets go ahead and clean up and make twilio live.

* Fixed "reset_cooldown" error.

reset_cooldown was giving an error because the cog didn't have reset_cooldown. It was supposed to be called inside the function ping since it has the decorator on it.

Changed so it calls reset_cooldown from the ping function and then changed all strings to use f-strings. Fixes #4

* Removed "live" progress. Changed all "mention" to "name".

Also removed progress for the live command. It just doesn't make any sense when they can just unload the cog. Doesn't make sense for development purpose either because if we're editing the file we can just comment out the offending lines anyway. Fixes #5

Grant was right. Using mentions is annoying.

* Fixed Syntax Error

Should be ready for pull now...