Pre-Production (#7) (#8)
* Change from loading txt to loading a python extension

We'll load from keys.py instead of keys.txt. End result is more efficient and cleaner looking.

README.md updated to help user with keys.py

Issues may arise from user injecting code into keys.py. This may be non-issue.

* Accidentally used my real token

Whoops

* Changed so the API reads twilio information now.

The entire thing actually works now.

* Cooldown (and extended commands) Initial Implementation (#2)

Reworked how the ping command works so now we are using the extended commands from discord.py. 

It's nice to import the extended commands instead of creating our own cooldown so we're not re-inventing the wheel. They also have their own error handling!

* Adding first command for ctx

Have to comment and rethink how we do the commands if we're going to use ctx.

* changed all events and calls to go through bot

bot now runs but does not function when !test is called. mention is a required argument that is missing.

* Changing to use 'Commands'

Revised code to ensure that the new ping command works while using ctx.

* Cleanup and added error handling for ping

removed time import because we will use the extended commands for that. Also ctx naturally handles multiple mentions so we will just make it so it only takes one. Changed format so it returns a mention instead of an id.

* Error handling now works for the ping command.

Instead of pulling the error exception inside the function we can use the event handling that ext has to only be checked when an actual error is called. In the documentation at ext the order of error and ctx was switched so it took a long time to find this. I also changed the format of some of our longer strings so they fit on half a screen because it looks cleaner. Pretty sure it's part of pep8 or something as well but we can worry about that later.

* Cooldown now functional.

Cooldown now functions as it is supposed to. Error handling provides good information to console. Removed debugging output and removed tsundere from the bot.

* Ping Error Handles User Not Found in Directory

Ping did not handle a user not being in the dictionary and put a cooldown on the user. We shouldn't punish the user from the command if it doesn't execute the command.

Found out how to use reset_cooldown. This can be abused to reset someone's cooldown if someone already has one on them though. We'll have to implement some kind of check for that. Maybe we can change this so an error is sent at command call instead.

* Handle No Arguments Provided.

manually throw an exception when we don't get a member.

I don't understand why this gets past the discord.Member conversion but we'll deal with it.

* Properly Handle No Arguments Provided.

I was wondering why it got past the discord.Member formatting and then realized that it looks like we set it to None after the formatting. That would explain why we get passed NoneType.

We'll avoid manually raising our own exception and then we can add it back into the error handling function. Added it to the BadArgument condition because that's really where it fits.

* Cogs Initial Implementation (#3)

* Initial

Implementing cogs so we are able to work in separate files and we can easily work on separate functions.

Doing it how we were currently doing it, SahibBot.py would have become a huge file and probably not very managable.

* Moved Error and Ping to their own cogs.

All of the ping and error handling are now in their own files and it works properly.

* Cleanup and made Twilio Live

We're able to comment out and unload extensions now so lets go ahead and clean up and make twilio live.

* Fixed "reset_cooldown" error. Mentions are now Name. Strings are now F-Strings (#6)

* Initial

Implementing cogs so we are able to work in separate files and we can easily work on separate functions.

Doing it how we were currently doing it, SahibBot.py would have become a huge file and probably not very managable.

* Moved Error and Ping to their own cogs.

All of the ping and error handling are now in their own files and it works properly.

* Cleanup and made Twilio Live

We're able to comment out and unload extensions now so lets go ahead and clean up and make twilio live.

* Fixed "reset_cooldown" error.

reset_cooldown was giving an error because the cog didn't have reset_cooldown. It was supposed to be called inside the function ping since it has the decorator on it.

Changed so it calls reset_cooldown from the ping function and then changed all strings to use f-strings. Fixes #4

* Removed "live" progress. Changed all "mention" to "name".

Also removed progress for the live command. It just doesn't make any sense when they can just unload the cog. Doesn't make sense for development purpose either because if we're editing the file we can just comment out the offending lines anyway. Fixes #5

Grant was right. Using mentions is annoying.

* Fixed Syntax Error

Should be ready for pull now...