Highlight (#12)
* Initial Commit

Sends a message to highlights on a react event.

* Embed Format and Reaction Count

Embeds are a great way to hold all the content without losing formatting like code blocks do.

We'll also display all the reactions inside of the highlight.

* Clarified Variable Names

Random one letter variables are not very descriptive. Was supposed to be in the last commit but I'm bad at hitting the save button.

* All functionality complete.

Highlight cog now works as expected.

* Bot Cannot Highlight Itself

Sure, it's not a big deal but I wouldn't expect that the bot would highlight itself. Plus, it'd be kind of nice to be able to react after the fact inside the highlight channel.

Added a simple ID check right before the on_reaction