Logging (#14)
We now have a logging function that can be optionally used in your cogs. It's as easy as importing logging and including log statements in your code. Amazing!

* Logging now works

Just have to add it to the other cogs so that they all log properly.

Ping is probably the best module to base logging properly in. We should probably talk about how much we want to log though.

* Logging Functions as Expected

I'm not so sure this can be created into it's own module but logging now works as you expect it to. We will still need to go over how much logging we want in future scripts as well as the format of how we want it.

The console now does not show anything besides saying that the bot is ready. All traces and errors SHOULD go to the log file.

This change makes it easier to see if there are any errors that happen in the program over time. We have time data as well as being able to search through to see how many times everything was called.

* Changed logging back to INFO